AND-8562 Make single DI module for all configs storages
